@@ -5,7 +5,7 @@ import { NgxFileDropEntry, NgxFileDropModule } from 'ngx-file-drop';
 import { MatButtonModule } from '@angular/material/button';
 import { VideoService } from '../services/video.service';
 import { CommonModule } from '@angular/common';
-import { HttpClientModule } from '@angular/common/http';
+import {} from '@angular/common/http';
 import { ToastrModule, ToastrService } from 'ngx-toastr';
 import { changeLoaderStatus } from '../shared/shared-function';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
@@ -16,7 +16,11 @@ import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
   imports: [
     FormsModule,
     CommonModule,
-    HttpClientModule,
+    
+// TODO: `HttpClientModule` should not be imported into a component directly.
+// Please refactor the code to add `provideHttpClient()` call to the provider list in the
+// application bootstrap logic and remove the `HttpClientModule` import from this component.
+HttpClientModule,
     NgxFileDropModule,
     MatButtonModule,
     ToastrModule,
